UI Changes for Someone Else/AddOn

+++ Enrich_SMA/Enrich_SMA/Modules/Appointment/Appointment Details/Client Information/ClientInformationVC.swift
@@ -35,7 +35,7 @@ class ClientInformationVC: UIViewController, ClientInformationDisplayLogic {
 
     @IBOutlet weak private var btnSubmitForm: UIButton!
 
-    var customerId: Int?
+    var customerId: Int64?
 
     var appointmentDetails: Appointment.GetAppointnents.Data?
